ingest/requirements.txt
requests
beautifulsoup4
google-cloud-storage
google-cloud-bigquery
